// app/careers/[jobId]/types.ts
export type ScreeningDetailsPublicResponseDTO = {
    title: string;
    company_name?: string | null;
    jd_url?: string | null;
    job_summary?: string | null;
    salary?: string | null;
    location?: string | null;
    employment_type?: string | null;
    work_arrangement?: string | null;
    department?: string | null;
    details?: string | null;
    posted_at?: string | null;
};